+++ src/app/sign-in/page.tsx
@@ -1,5 +1,9 @@
-import SignInPage from '@/pages/SignInPage';
+import SignInModal from '@/features/sign-in/components/SignInModal';
 
 export default function WrapperSignInPage() {
-  return <SignInPage />;
+  return (
+    <div className="bg-background h-screen">
+      <SignInModal />
+    </div>
+  );
 }
